ANDROID: cpufreq: change uid_cpupower to use u32
The uid_cpupower files use u64 to report how long uid's have spent
on the cpu. Using u64 instead of u32 has a noticable impact on
the time taken to read a file. These file reads are on a critical
path to turning the screen on.

Using u64:
/proc/uid_cpupower/time_in_state
0.00584s real
0.0000134s user
0.00579s system

/proc/uid_cpupower/concurrent_active_time
0.00243s real
0.00000335s user
0.00241s system

/proc/uid_cpupower/concurrent_policy_time
0.00244s real
0.00000320s user
0.00243s system

Using u32:
/proc/uid_cpupower/time_in_state
0.00358s real
0.00000488s user
0.00356s system

/proc/uid_cpupower/concurrent_active_time
0.000851s real
0.00000274s user
0.000843s system

/proc/uid_cpupower/concurrent_policy_time
0.000824s real
0.00000244s user
0.000817s system

When using u32 the max time that can be reported is 497 days.
Since it is unlikely that a uid has run for that long without
the phone rebooting or the uid dying, the performance gains make
switching to u32 worth it. In the worst case, the non-essential
battery blaming will be inaccurate for that uid.

This patch does not change the internal storage size or how other
related stats are reported. It only changes how uid_cpupower
reports the times.

b/72763654
Test: measure read times and verify file contents is correct

+++ drivers/cpufreq/cpufreq_stats.c
@@ -318,7 +318,7 @@ static int uid_time_in_state_show(struct seq_file *m, void *v)
 }
 
 /*
- * time_in_state is an array of u64's in the following format:
+ * time_in_state is an array of u32's in the following format:
  * [n, uid0, time0a, time0b, ..., time0n,
  *     uid1, time1a, time1b, ..., time1n,
  *     uid2, time2a, time2b, ..., time2n, etc.]
@@ -328,8 +328,8 @@ static int time_in_state_show(struct seq_file *m, void *v)
 {
 	struct uid_entry *uid_entry;
 	unsigned long bkt;
-	u64 cpufreq_max_states = all_freq_table->table_size;
-	u64 uid, time;
+	u32 cpufreq_max_states = all_freq_table->table_size;
+	u32 uid, time;
 	int i;
 
 	if (!cpufreq_all_freq_init)
@@ -348,12 +348,12 @@ static int time_in_state_show(struct seq_file *m, void *v)
 			max_state = uid_entry->alive_max_states;
 
 		if (max_state) {
-			uid = (u64) uid_entry->uid;
+			uid = (u32) uid_entry->uid;
 			seq_write(m, &uid, sizeof(uid));
 		}
 
 		for (i = 0; i < max_state; ++i) {
-			u64 total_time_in_state = 0;
+			u32 total_time_in_state = 0;
 
 			if (uid_entry->dead_time_in_state &&
 				i < uid_entry->dead_max_states) {
@@ -365,7 +365,7 @@ static int time_in_state_show(struct seq_file *m, void *v)
 				total_time_in_state +=
 					uid_entry->alive_time_in_state[i];
 			}
-			time = (u64)
+			time = (u32)
 				cputime_to_clock_t(total_time_in_state);
 			seq_write(m, &time, sizeof(time));
 		}
@@ -380,7 +380,7 @@ static int time_in_state_show(struct seq_file *m, void *v)
 }
 
 /*
- * concurrent_active_time is an array of u64's in the following format:
+ * concurrent_active_time is an array of u32's in the following format:
  * [n, uid0, time0a, time0b, ..., time0n,
  *     uid1, time1a, time1b, ..., time1n,
  *     uid2, time2a, time2b, ..., time2n, etc.]
@@ -390,7 +390,7 @@ static int concurrent_active_time_show(struct seq_file *m, void *v)
 {
 	struct uid_entry *uid_entry;
 	struct task_struct *task, *temp;
-	u64 uid, time, num_possible_cpus = num_possible_cpus();
+	u32 uid, time, num_possible_cpus = num_possible_cpus();
 	unsigned long bkt, flags;
 	int i;
 
@@ -434,11 +434,11 @@ static int concurrent_active_time_show(struct seq_file *m, void *v)
 			!uid_entry->dead_concurrent_active_time)
 			continue;
 
-		uid = (u64) uid_entry->uid;
+		uid = (u32) uid_entry->uid;
 		seq_write(m, &uid, sizeof(uid));
 
 		for (i = 0; i < num_possible_cpus; ++i) {
-			u64 total_concurrent_active_time = 0;
+			u32 total_concurrent_active_time = 0;
 
 			if (uid_entry->dead_concurrent_active_time)
 				total_concurrent_active_time =
@@ -446,7 +446,7 @@ static int concurrent_active_time_show(struct seq_file *m, void *v)
 			if (uid_entry->alive_concurrent_active_time)
 				total_concurrent_active_time +=
 					uid_entry->alive_concurrent_active_time[i];
-			time = (u64)
+			time = (u32)
 				cputime_to_clock_t(total_concurrent_active_time);
 			seq_write(m, &time, sizeof(time));
 		}
@@ -462,7 +462,7 @@ static int concurrent_active_time_show(struct seq_file *m, void *v)
 }
 
 /*
- * concurrent_policy_time is an array of u64's in the following format:
+ * concurrent_policy_time is an array of u32's in the following format:
  * [n, x0, ..., xn, uid0, time0a, time0b, ..., time0n,
  *                  uid1, time1a, time1b, ..., time1n,
  *                  uid2, time2a, time2b, ..., time2n, etc.]
@@ -475,8 +475,8 @@ static int concurrent_policy_time_show(struct seq_file *m, void *v)
 	struct task_struct *task, *temp;
 	struct cpufreq_policy *policy;
 	struct cpufreq_policy *last_policy = NULL;
-	u64 buf[num_possible_cpus()];
-	u64 uid, time;
+	u32 buf[num_possible_cpus()];
+	u32 uid, time;
 	unsigned long bkt, flags;
 	int i, cnt = 0, num_possible_cpus = num_possible_cpus();
 
@@ -501,7 +501,7 @@ static int concurrent_policy_time_show(struct seq_file *m, void *v)
 	if (last_policy)
 		cpufreq_cpu_put(last_policy);
 
-	buf[0] = (u64) cnt;
+	buf[0] = (u32) cnt;
 	seq_write(m, buf, (cnt + 1) * sizeof(*buf));
 
 	rt_mutex_lock(&uid_lock);
@@ -539,11 +539,11 @@ static int concurrent_policy_time_show(struct seq_file *m, void *v)
 			!uid_entry->dead_concurrent_policy_time)
 			continue;
 
-		uid = (u64) uid_entry->uid;
+		uid = (u32) uid_entry->uid;
 		seq_write(m, &uid, sizeof(uid));
 
 		for (i = 0; i < num_possible_cpus; ++i) {
-			u64 total_concurrent_policy_time = 0;
+			u32 total_concurrent_policy_time = 0;
 
 			if (uid_entry->dead_concurrent_policy_time)
 				total_concurrent_policy_time =
@@ -551,7 +551,7 @@ static int concurrent_policy_time_show(struct seq_file *m, void *v)
 			if (uid_entry->alive_concurrent_policy_time)
 				total_concurrent_policy_time +=
 					uid_entry->alive_concurrent_policy_time[i];
-			time = (u64)
+			time = (u32)
 				cputime_to_clock_t(total_concurrent_policy_time);
 			seq_write(m, &time, sizeof(time));
 		}
